fixed alloc bug in tkidx (cf #2371) and msvc warnings

+++ src/utils/url.c
@@ -314,9 +314,9 @@ static char *gf_url_concatenate_ex(const char *parentName, const char *pathName,
 	if (relative_to_parent) {
 		u32 l1=0, l2=0;
 		rad = strrchr(tmp, '/');
-		if (rad) l1 = rad - tmp;
+		if (rad) l1 = (u32) (rad - tmp);
 		rad = strrchr(pathName, '/');
-		if (rad) l2 = rad - pathName;
+		if (rad) l2 = (u32) (rad - pathName);
 		if (l1 && l2 && (l1==l2) && !strncmp(pathName, parentName, l1)) {
 			outPath = gf_strdup(pathName + l1 + 1);
 			goto check_spaces;
